move activation of kontainer to Rancher
